Optimize matrix operations.
Reducing the matrices to common verses is unneccessary. The matrix
multiplication afterwards will see to that, too. That does not change
the performance characteristics. But it enables the next step:
Transposing a COO matrix is faster than operating on a CSC matrix.
So I constructed the CSC matrix from the transposed COO matrix.